fix: eslint warning
