+++ rs-react-app/src/other/main.tsx
@@ -1,0 +1,5 @@
+const Main = () => {
+  return null;
+};
+
+export default Main;
